+++ src/test/OpenEXRTest/testPreviewImage.cpp
@@ -121,13 +121,13 @@ readWriteFiles (const char fileName1[],
 	file2.setFrameBuffer (pixels2 - dx - dy * w, 1, w);
 	file2.readPixels (dw.min.y, dw.max.y);
 
-	for (size_t i = 0; i < w * h; ++i)
-	{
-	    assert (pixels1[i].r == pixels2[i].r);
-	    assert (pixels1[i].g == pixels2[i].g);
-	    assert (pixels1[i].b == pixels2[i].b);
-	    assert (pixels1[i].a == pixels2[i].a);
-	}
+        for (int i = 0; i < w * h; ++i)
+        {
+            assert (pixels1[i].r == pixels2[i].r);
+            assert (pixels1[i].g == pixels2[i].g);
+            assert (pixels1[i].b == pixels2[i].b);
+            assert (pixels1[i].a == pixels2[i].a);
+        }
     }
 
     cout << "writing file " << fileName3 << endl;
